writing up webcam recorder program
