+++ pom.xml
@@ -38,6 +38,12 @@
             <artifactId>javacv-platform</artifactId>
             <version>1.4.2</version>
         </dependency>
+
+        <dependency>
+            <groupId>org.bytedeco.javacpp-presets</groupId>
+            <artifactId>flycapture-platform</artifactId>
+            <version>2.11.3.121-1.4.2</version>
+        </dependency>
         <!-- https://mvnrepository.com/artifact/org.apache.commons/commons-csv -->
         <dependency>
             <groupId>org.apache.commons</groupId>
